requirements.txt
python-dotenv==1.0.1
pyspark==3.5.1
numpy==1.26.4
# pandas==2.3.0
pandas==2.2.2
dbt-core==1.9.2
dbt-snowflake==1.9.2
snowflake-connector-python==3.15.0
protobuf==5.29.0
cffi==1.15.1
cryptography==41.0.7
requests==2.32.4
elementary-data==0.19.1
scikit-learn==1.4.2
joblib==1.4.2